+++ test/sql/adm3.sql
@@ -563,6 +563,7 @@ select hist_function, hist_event, hist_object,
   from emaj.emaj_hist where hist_id >= 18000 order by hist_id;
 
 select * from emaj.emaj_log_session where lower(lses_time_range) >= 12000 order by lses_group, lses_time_range;
+select * from emaj.emaj_group_hist order by grph_group, grph_time_range;
 
 -----------------------------
 -- check grants on other functions to emaj_adm role