+++ CHANGES.md
@@ -3,6 +3,7 @@ E-Maj - Change log
 \<devel\>
 ------
 ### Enhancements: ###
+  * Add an internal table that records group creations and drops.
   * Add an internal table that records log sessions, i.e. time ranges between
     start_group and stop_group operations.
   * When computing log statistics or generating SQL statements for changes